+++ MTR Lineages/scripts/mtrLineages/lineages.lua
@@ -1,5 +1,6 @@
 local I = require("openmw.interfaces")
 local self = require("openmw.self")
+local core = require("openmw.core")
 
 local selfSkills = self.type.stats.skills
 local selfAttrs = self.type.stats.attributes
@@ -77,6 +78,61 @@ local function getRaceId(npc)
     ---@diagnostic disable-next-line: undefined-field
     return npc.type.records[npc.recordId].race
 end
+
+---@diagnostic disable: duplicate-doc-field
+
+---@class Dependency
+---@field plugin      string   esp/omwaddon/omwscripts filename of the required plugin
+---@field interface   any      The interface object retrieved from the other mod
+---@field minVersion  number|nil
+---@field curVersion  number|nil
+
+---@param dep  Dependency
+---@return     boolean, string|nil
+local function checkDependency(dep)
+    local checks = {
+        {
+            ok  = core.contentFiles.has(dep.plugin:lower()),
+            msg = ("'%s' dependency not found."):format(dep.plugin)
+        },
+        {
+            ok  = dep.interface ~= nil,
+            msg = ("'%s' has to be loaded before this mod."):format(dep.plugin)
+        },
+        {
+            ok  = not dep.minVersion or dep.curVersion >= dep.minVersion,
+            msg = ("'%s' version too low. Required %s, found %s.")
+                :format(dep.plugin, tostring(dep.minVersion), tostring(dep.curVersion))
+        },
+    }
+    for _, c in ipairs(checks) do
+        if not c.ok then
+            return false, c.msg
+        end
+    end
+    return true
+end
+
+---@param modName  string
+---@param depList  Dependency[]
+local function checkAll(modName, depList)
+    local errors = {}
+    for _, dep in ipairs(depList) do
+        local ok, msg = checkDependency(dep)
+        if not ok then
+            errors[#errors + 1] = msg
+        end
+    end
+    if #errors > 0 then
+        local msg = ("[%s]\nDependency error.\n\n%s\n"):format(modName, table.concat(errors, "\n\n"))
+        I.UI.showInteractiveMessage(msg)
+    end
+end
+
+checkAll("MTR Lineages", {
+    plugin = "CharacterTraitsFramework.omwscripts",
+    interface = I.CharacterTraits,
+})
 
 I.CharacterTraits.addTrait {
     id = "spriggan",
@@ -1758,7 +1814,7 @@ I.CharacterTraits.addTrait {
         selfSpells:add("mtrLineage_Breton")
 
         attrs.endurance.base = attrs.endurance.base - 5
-        attrs.personastrengthlity.base = attrs.strength.base - 5
+        attrs.strength.base = attrs.strength.base - 5
         attrs.agility.base = attrs.agility.base - 5
         attrs.speed.base = attrs.speed.base - 5
     end,
@@ -1793,7 +1849,7 @@ I.CharacterTraits.addTrait {
         selfSpells:add("mtrLineage_Breton")
 
         attrs.endurance.base = attrs.endurance.base - 5
-        attrs.personastrengthlity.base = attrs.strength.base - 5
+        attrs.strength.base = attrs.strength.base - 5
         attrs.agility.base = attrs.agility.base - 5
         attrs.speed.base = attrs.speed.base - 5
     end,